script renaming part 1
